DPC-678: Refactor token operations into API service
This is just a lift and shift, no tests have been run yet.
