+++ dpc-api/pom.xml
@@ -72,6 +72,12 @@
             <scope>compile</scope>
         </dependency>
         <dependency>
+            <groupId>gov.cms.dpc</groupId>
+            <artifactId>dpc-macaroons</artifactId>
+            <version>0.4.0-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
             <groupId>${hapi.fhir.groupID}</groupId>
             <artifactId>hapi-fhir-client</artifactId>
         </dependency>
