+++ dpc-api/src/main/java/gov/cms/dpc/api/auth/AuthModule.java
@@ -4,6 +4,8 @@ import com.google.inject.Binder;
 import com.google.inject.TypeLiteral;
 import com.hubspot.dropwizard.guicier.DropwizardAwareModule;
 import gov.cms.dpc.api.DPCAPIConfiguration;
+import gov.cms.dpc.api.auth.macaroons.BakeryProvider;
+import gov.cms.dpc.macaroons.MacaroonBakery;
 import io.dropwizard.auth.Authenticator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -35,5 +37,6 @@ public class AuthModule extends DropwizardAwareModule<DPCAPIConfiguration> {
             binder.bind(authenticatorTypeLiteral).to(MacaroonsAuthenticator.class);
         }
         binder.bind(DPCAuthDynamicFeature.class);
+        binder.bind(MacaroonBakery.class).toProvider(BakeryProvider.class);
     }
 }
