+++ dpc-api/src/main/java/gov/cms/dpc/api/auth/macaroons/BakeryProvider.java
@@ -1,7 +1,8 @@
-package gov.cms.dpc.attribution.macaroons;
+package gov.cms.dpc.api.auth.macaroons;
 
-import gov.cms.dpc.attribution.DPCAttributionConfiguration;
-import gov.cms.dpc.attribution.config.TokenPolicy;
+import gov.cms.dpc.api.DPCAPIConfiguration;
+import gov.cms.dpc.api.config.TokenPolicy;
+import gov.cms.dpc.common.annotations.APIV1;
 import gov.cms.dpc.macaroons.MacaroonBakery;
 import gov.cms.dpc.macaroons.store.IRootKeyStore;
 import gov.cms.dpc.macaroons.thirdparty.IThirdPartyKeyStore;
@@ -16,21 +17,23 @@ import javax.inject.Singleton;
 @Singleton
 public class BakeryProvider implements Provider<MacaroonBakery> {
 
-    private final DPCAttributionConfiguration configuration;
+    private final DPCAPIConfiguration configuration;
     private final IRootKeyStore store;
     private final IThirdPartyKeyStore thirdPartyKeyStore;
+    private final String publicURL;
 
     @Inject
-    public BakeryProvider(DPCAttributionConfiguration configuration, IRootKeyStore store, IThirdPartyKeyStore thirdPartyKeyStore) {
+    public BakeryProvider(DPCAPIConfiguration configuration, IRootKeyStore store, IThirdPartyKeyStore thirdPartyKeyStore, @APIV1 String publicURI) {
         this.configuration = configuration;
         this.store = store;
         this.thirdPartyKeyStore = thirdPartyKeyStore;
+        this.publicURL = publicURI;
     }
 
     @Override
     public MacaroonBakery get() {
         final TokenPolicy tokenPolicy = this.configuration.getTokenPolicy();
-        return new MacaroonBakery.MacaroonBakeryBuilder(configuration.getPublicServerURL(), store, thirdPartyKeyStore)
+        return new MacaroonBakery.MacaroonBakeryBuilder(publicURL, store, thirdPartyKeyStore)
                 .addDefaultCaveatSupplier(new VersionCaveatSupplier(tokenPolicy))
                 .addDefaultCaveatSupplier(new ExpirationCaveatSupplier(tokenPolicy))
                 .addDefaultVerifier(new VersionCaveatVerifier(tokenPolicy))
