+++ dpc-api/src/main/java/gov/cms/dpc/api/auth/macaroons/ExpirationCaveatVerifier.java
@@ -1,8 +1,7 @@
-package gov.cms.dpc.attribution.macaroons;
+package gov.cms.dpc.api.auth.macaroons;
 
-import gov.cms.dpc.attribution.config.TokenPolicy;
+import gov.cms.dpc.api.config.TokenPolicy;
 import gov.cms.dpc.macaroons.CaveatVerifier;
-import gov.cms.dpc.macaroons.MacaroonCaveat;
 import gov.cms.dpc.macaroons.MacaroonCondition;
 
 import java.time.OffsetDateTime;
