+++ dpc-api/src/main/java/gov/cms/dpc/api/auth/macaroons/VerifierConstants.java
@@ -1,4 +1,4 @@
-package gov.cms.dpc.attribution.macaroons;
+package gov.cms.dpc.api.auth.macaroons;
 
 public class VerifierConstants {
