+++ dpc-api/src/main/java/gov/cms/dpc/api/auth/macaroons/VersionCaveatVerifier.java
@@ -1,13 +1,12 @@
-package gov.cms.dpc.attribution.macaroons;
+package gov.cms.dpc.api.auth.macaroons;
 
-import gov.cms.dpc.attribution.config.TokenPolicy;
+import gov.cms.dpc.api.config.TokenPolicy;
 import gov.cms.dpc.macaroons.CaveatVerifier;
-import gov.cms.dpc.macaroons.MacaroonCaveat;
 import gov.cms.dpc.macaroons.MacaroonCondition;
 
 import java.util.Optional;
 
-import static gov.cms.dpc.attribution.macaroons.VerifierConstants.NO_MATCH;
+import static gov.cms.dpc.api.auth.macaroons.VerifierConstants.NO_MATCH;
 
 /**
  * Implementation of {@link CaveatVerifier} which verifies that token version meets a minimum threshold
