+++ dpc-attribution/src/main/java/gov/cms/dpc/attribution/resources/AbstractAttributionResource.java
@@ -28,9 +28,6 @@ public abstract class AbstractAttributionResource {
     @Path("/Practitioner")
     public abstract AbstractPractitionerResource providerOperations();
 
-    @Path("/Token")
-    public abstract AbstractTokenResource tokenOperations();
-
     @GET
     @Path("/_healthy")
     @ApiOperation(value = "Check is healthy", notes = "Returns whether or not the application is in a healthy state." +
