+++ dpc-attribution/src/main/java/gov/cms/dpc/attribution/resources/v1/V1AttributionResource.java
@@ -13,21 +13,18 @@ public class V1AttributionResource extends AbstractAttributionResource {
     private final EndpointResource er;
     private final PatientResource par;
     private final PractitionerResource pr;
-    private final TokenResource tr;
 
     @Inject
     public V1AttributionResource(GroupResource gr,
                                  OrganizationResource or,
                                  EndpointResource er,
                                  PatientResource par,
-                                 PractitionerResource pr,
-                                 TokenResource tr) {
+                                 PractitionerResource pr) {
         this.gr = gr;
         this.or = or;
         this.er = er;
         this.par = par;
         this.pr = pr;
-        this.tr = tr;
     }
 
     @Override
@@ -53,10 +50,5 @@ public class V1AttributionResource extends AbstractAttributionResource {
     @Override
     public AbstractPractitionerResource providerOperations() {
         return this.pr;
-    }
-
-    @Override
-    public AbstractTokenResource tokenOperations() {
-        return this.tr;
     }
 }
